+++ NEWS.md
@@ -1,3 +1,5 @@
+# tidypredict (development version)
+
 # tidypredict 0.3.0
 
 ## New features