+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## New features
 
+- Adds support for `parsnip` fitted models: `lm`, `randomForest`, `ranger`, and `earth`
+
 - Adds support for xgb.Booster models provided by the `xgboost` package (@Athospd, #43)
 
 - Adds support for `Cubist::cubist()` models